GP-71 change to useContext and write test for it.
